v 0.1.5661.33254
The new format for urls as they are written to file is: "*URL*
*RESOLUTION* *FORMAT*". For anyone upgrading, you need to put the format
for each url after it's resolution, or the program will crash.

Otherwise, not much was done, additional resolutions are available for
every resolution beside Mp4, which remains locked at 360. Obviously
additional format support is now included, and hopefully scheduling will
be available soon.