Update 3
Releasing a little early due to the amount of fixes that need to be done
to only slightly optimize the program....

Update 4 will literally be a conversion from WinForms to WPF, with bug
fixes where applicable. It is very unlikely a feature will be added, but
I don't discount the possibility.

Full details contained in dev notes.