Update 5.1
Due to a major oversight on my part, the audio downloading feature was
completely unimplemented, as well as a few other bugs I missed which
made downloading confusing. Because of this, I upload this "patch". As
far as I know, everything that is supposed to work does, except for a
few things mentioned in the update notes.