Update 6.35
A small update to get rid of some bugs in the download process which
made the program a pain to use. Because it can take me a while to
release a fully functioning new update, I felt it a little selfish to
make you all deal with the problems while I don't have to.