Update 7.5 FINAL
I am making the move from C# to Python for all code. Meaning this
version of the program is now defunct.

I will leave this version of the program up for posterity as well as a
buffer between these new versions as I am just learning Python and it
will take me a while to get comfortable with it.

While this is a development version not intended for stable release it
appears to be stable regardless, so I am comfortable releasing it this
way.